scripts/TextureEditor/requirements.txt
tkcolorpicker
numpy
opencv-python
Pillow